refactor: move trend calculation function to charts module for better organization
